% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R Functions Oxyzen.R
\name{E.P.Leakage.trbb}
\alias{E.P.Leakage.trbb}
\title{Expected Leakage and Probability from Truncated Beta Distribution}
\usage{
E.P.Leakage.trbb(alpha, beta, cutoff, B, b, m, M, Delta = 1)
}
\arguments{
\item{alpha}{Numeric vector of shape parameters for the Beta distribution.}

\item{beta}{Numeric vector of shape parameters for the Beta distribution.}

\item{cutoff}{Numeric value in (0, 1). The truncation point for the Beta distribution.}

\item{B}{Integer. Total number of groups (say, bags) in a batch.}

\item{b}{Integer. Sampled number of groups (say, sample bags used in group-test) for a batch.}

\item{m}{Integer. Pool-size, i.e, number of sampled units used in group-test from each sampled group.}

\item{M}{Integer. Total number of units in a group (bag).}

\item{Delta}{Numeric scalar for sensitivity. Scale factor to adjust sensitivity (default = 1).}
}
\value{
A data frame with columns:
\describe{
  \item{alpha}{Input vector of alpha values.}
  \item{beta}{Input vector of beta values.}
  \item{E_Li}{Expected leakage for each (alpha, beta) pair.}
  \item{P_Li}{Estimated probability of leakage.}
}
}
\description{
Computes the expected leakage and the probability of leakage from a  formulation using
a truncated Beta distribution. The function provides values from analytic forms and uses log-scale incomplete
Beta functions to avoid underflow. Analytic function are based on the perfect specificity. The probability of leakage is estimated using an approximate distribution for the group probability of being contaminated when sensitivity is imperfect. For perfect sensitivity, there is no approximate distribution required.
}
\details{
This function evaluates integrals over the tail of the Beta distribution (i.e., from `cutoff` to 1)
using log-scale calculations to maintain numerical accuracy. The expected leakage is calculated as a
ratio of incomplete Beta integrals, weighted by group sizes. Two formulations for probability of leakage are used depending
on whether `Delta` equals 1 (exact) or differs from 1 (approximation).
}
\examples{
alpha_vals <- c(0.004, 0.005)
beta_vals <- c(5.0, 6.0)
result <- E.P.Leakage.trbb.semi.exact(
  alpha = alpha_vals,
  beta = beta_vals,
  cutoff = 0.02,
  B = 8000, b = 13,
  m = 5, M = 40,
  Delta = 0.70
)
print(result)

}
