% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_algorithm.R
\name{inverse_logit}
\alias{inverse_logit}
\title{Compute the Inverse Logit (Logistic Function)}
\usage{
inverse_logit(x)
}
\arguments{
\item{x}{A numeric value or vector. Typically a log-odds or any real number.}
}
\value{
A numeric value (or vector) representing the inverse logit of the input.
}
\description{
This function computes the inverse logit transformation, also known as the logistic function.
It maps real-valued inputs to the (0, 1) interval.
}
\examples{
inverse_logit(0.8472979)  # Approximately 0.7
inverse_logit(c(-1, 0, 1))

}
