% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_algorithm.R
\name{logit}
\alias{logit}
\title{Compute the Logit of a Probability}
\usage{
logit(p)
}
\arguments{
\item{p}{A numeric value or vector strictly between 0 and 1. Represents the probability.}
}
\value{
A numeric value (or vector) representing the logit of the input probability.
}
\description{
This function computes the logit (log-odds) transformation of a probability value.
}
\examples{
logit(0.7)
logit(c(0.3, 0.5, 0.7))

}
