% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_algorithm.R
\name{rtrbeta}
\alias{rtrbeta}
\title{Generate Random Samples from a Truncated Beta Distribution}
\usage{
rtrbeta(n, shape1, shape2, cutoff)
}
\arguments{
\item{n}{Integer. Number of samples to draw.}

\item{shape1}{Numeric. First shape parameter of the Beta distribution.}

\item{shape2}{Numeric. Second shape parameter of the Beta distribution.}

\item{cutoff}{Numeric between 0 and 1. Values sampled below this threshold are set to 0.}
}
\value{
A numeric vector of length \code{n}, containing truncated Beta samples.
}
\description{
Draws random samples from a Beta distribution and truncates them at a specified cutoff value.
If a sampled value is less than the cutoff, it is replaced with 0.
}
\examples{
set.seed(123)
rtrbeta(10, shape1 = 2, shape2 = 5, cutoff = 0.2)

}
