% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE Functions.R
\name{fit_GroupedBB}
\alias{fit_GroupedBB}
\title{Fit Grouped Beta-Binomial (GroupedBB) Model to Grouped Data}
\usage{
fit_GroupedBB(
  ty,
  freq,
  b,
  m,
  G = 10000,
  M = NULL,
  B = NULL,
  sensitivity = 1,
  specificity = 1,
  theta = Inf,
  control = list(reltol = 1e-12, fnscale = -1),
  hessian = FALSE,
  leakage = FALSE
)
}
\arguments{
\item{ty}{Numeric vector. A numeric vector of grouped test outcomes (e.g., number of test positive groups).}

\item{freq}{Numeric vector. Frequencies corresponding to each count in `ty`.}

\item{b}{Integer. Number of groups tested per replicate (i.e., binomial trials).}

\item{m}{Integer. Number of individuals included in each pooled group test.}

\item{G}{Integer. Number of grid points from the beta distribution used to approximate the integral (default = 1000).}

\item{M}{Integer, optional. Required if leakage = TRUE. Total number of individuals in the larger group.
The value of `m` can be equal to M if all the group members are used for testing.}

\item{B}{Integer, optional. Required if leakage = TRUE. Total number of groups in a replicate (e.g., batch).}

\item{sensitivity}{Numeric, default 1. Sensitivity parameter for misclassification or detection.}

\item{specificity}{Numeric, default 1. Specificity parameter for misclassification or detection.}

\item{theta}{Clustering parameter. Use `Inf` for considering random groups.}

\item{control}{List. Control parameters passed to `optim()`.}

\item{hessian}{Logical, default FALSE. Whether to return the Hessian matrix from optimization.}

\item{leakage}{Logical, default FALSE. If TRUE (requires `M` and `B`), compute leakage quantities such as probability (`P.L`) and expected leakage (`E.L`) assuming perfect specificity.}
}
\value{
A list containing estimated parameters `alpha`, `beta`, `mu` (mean parameter),
        sensitivity, specificity, log-likelihood value (`loglik`), and convergence status.
        If `hessian = TRUE`, also returns the Hessian matrix.
        If `leakage = TRUE`, includes leakage-related estimates `P.L` and `E.L`.
}
\description{
Fits a beta-binomial model to grouped count data using maximum likelihood estimation.
Estimates parameters alpha and beta unless provided, and supports leakage estimation
when `leakage = TRUE`.
}
\examples{
freq <- c(2815, 9, 10, 6, 1, 3, 2, 0, 1, 2, 1, 0, 0, 0)
ty <- 0:13

# Fit model without leakage
bb.m1 <- fit_GroupedBB(ty = ty, freq = freq, b = 13, m = 5, M = 40,
                       sensitivity = 1, specificity = 1)

# Fit model with leakage (requires B)
bb.m2 <- fit_GroupedBB(ty = ty, freq = freq, b = 13, m = 5, M = 40, B = 8000,
                       sensitivity = 1, specificity = 1, leakage = TRUE)

}
