% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_algorithm.R
\name{MH_Sampler_BB}
\alias{MH_Sampler_BB}
\title{Metropolis-Hastings Sampler with Truncated Beta Distribution}
\usage{
MH_Sampler_BB(
  par,
  initial,
  ty,
  wt,
  trail,
  group.size,
  cutoff,
  G,
  R,
  sigma,
  alpha,
  beta,
  mu,
  rho,
  num_chains = 3,
  burnin,
  thin = 5,
  seed,
  sensitivity,
  specificity,
  sensitivity.range = NULL,
  specificity.range = NULL
)
}
\arguments{
\item{par}{Initial parameter values (numeric vector).}

\item{initial}{Optional. Starting values for the parameters (numeric vector).}

\item{ty}{Integer vector of observed successes (e.g., number of groups tested positive per replicate).}

\item{wt}{Numeric vector. Frequencies corresponding to each count in \code{ty}.}

\item{trail}{Integer. Number of trials (e.g., number of group-tests conducted per batch).}

\item{group.size}{Integer. Number of individuals tested per group (pool size).}

\item{cutoff}{Truncation threshold for the Beta distribution (numeric).
cutoff = 0 leads to the simple beta-binomial model.}

\item{G}{Integer. Number of quantiles from the beta distribution required for integration.}

\item{R}{Integer. Number of MCMC iterations.}

\item{sigma}{Proposal standard deviation for each parameter (numeric vector).}

\item{alpha}{Logical. Whether to include \code{alpha} in the model.}

\item{beta}{Logical. Whether to include \code{beta} in the model.}

\item{mu}{Logical. Whether to include \code{mu} in the model.}

\item{rho}{Logical. Whether to include \code{rho} in the model.}

\item{num_chains}{Number of MCMC chains to run (default is 3).}

\item{burnin}{Number of initial iterations to discard (default is \code{R * 0.5}).}

\item{thin}{Thinning interval for MCMC samples (default is 5).}

\item{seed}{A vector of integers for random number seeds (one per chain).}

\item{sensitivity}{Logical. Whether to consider test sensitivity in the model.}

\item{specificity}{Logical. Whether to consider test specificity in the model.}

\item{sensitivity.range}{Optional, numeric vector. If sensitivity=TRUE, a numeric vector of length 2 indicating \link{min, max} for sensitivity
(e.g., sensitivity ranges over U(0.65, 0.75)).}

\item{specificity.range}{Optional, numeric vector. If specificity=TRUE, a numeric vector of length 2 indicating \link{min, max} for specificity
(e.g., specificity ranges over U(0.90, 0.99)).}

\item{M}{Integer, optional. Required if leakage = TRUE. Total number of individuals in the larger group.
The value of \code{m} can be equal to M if all the group members are used for testing.}
}
\value{
A list with four elements:
\describe{
\item{target.parameters}{List of posterior samples for each parameter across chains.}
\item{parameters}{List of MCMC samples for each parameter under the transformed scale using in the sampler across each chain.}
\item{log.likelihood.values}{List of log-likelihood values for each iteration.}
\item{log.likelihood.yi}{List of log-likelihood values for each observation of \code{ty} at each iteration.}
}
}
\description{
Runs a Metropolis–Hastings (MH) MCMC algorithm to sample from a posterior
distribution involving a truncated Beta distribution. The function supports
several parameterizations, and optionally models test sensitivity and specificity
with prior distributions.
}
\details{
Implements a flexible Bayesian MCMC sampler with support for multiple parameterizations
(via alpha, beta, mu, and rho). It includes likelihood computation using truncated
beta-distributed prevalence and allows modeling imperfect test sensitivity/specificity
using them in logit scales. The truncated beta-binomial model with cutoff k = 0
reduces to a standard beta-binomial model.
}
\examples{
\dontrun{
result <- MH_Sampler_BB(
  par = c(log(0.005), logit(7e-04)),
  cutoff = 0.01,
  alpha = TRUE, beta = FALSE, mu = TRUE, rho = FALSE,
  G = 500, R = 1000,
  sigma = c(0.20, 0.20),
  num_chains = 4,
  burnin = 200, thin = 5,
  seed = c(123, 456, 789, 102),
  trail = 13,
  ty = 0:13,
  wt = c(2815,9,10,6,1,3,2,0,1,2,1,0,0,0),
  group.size = 5,
  sensitivity = FALSE,
  specificity = FALSE
)
}

}
